// src/_workspace/types/remain-leave/RemainLeaveInterface.ts
export interface RemainLeaveInterface {
  EMPLOYEE_CODE: string
  EMPLOYEE_NAME: string
  EMPLOYEE_SURNAME: string
  EMPLOYEE_SECTION: string
  EMPLOYEE_START_WORK?: string
  SECTION?: string
  DEPARTMENT?: string
  START_WORK?: string
  LEAVE_TYPE?: string
  LEAVE_TYPE_ID: number
  LEAVE_TYPE_CODE?: string
  LEAVE_TYPE_DESCRIPTION_EN?: string
  LEAVE_TYPE_DESCRIPTION_TH?: string
  LEAVE_REMAIN_DAY?: number
  LEAVE_REMAINING?: number
  CREATE_BY?: string
  CREATE_DATE?: string
  UPDATE_BY?: string
  UPDATE_DATE?: string
  DESCRIPTION?: string
  INUSE?: string
}